+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Not released
 
+- cache Intl instances to improve performance [#894](https://github.com/CartoDB/carto-react/pull/894)
+
 ## 3.0.0
 
 ### 3.0.0-alpha.16 (2024-07-26)